requirements.txt
python-dotenv
click
